+++ manager-api/src/api/health/health.controller.ts
@@ -1,7 +1,9 @@
 import { Controller, Get } from "@nestjs/common";
+import { Public } from "../../core/auth/auth.decorator";
 import { HealthcheckService } from "../../core/healthcheck/healthcheck.service";
 import { HealthApi, HealthStatusDocs } from "./health.openapi";
 
+@Public()
 @HealthApi()
 @Controller({ path: "health", version: "1" })
 export class HealthController {