+++ manager-api/src/core/auth/api-token/api-token.module.ts
@@ -1,0 +1,13 @@
+import { Module } from "@nestjs/common";
+import { TypeOrmModule } from "@nestjs/typeorm";
+import { ApiTokenController } from "./api-token.controller";
+import { ApiToken } from "./api-token.entity";
+import { ApiTokenService } from "./api-token.service";
+
+@Module({
+  imports: [TypeOrmModule.forFeature([ApiToken])],
+  providers: [ApiTokenService],
+  controllers: [ApiTokenController],
+  exports: [ApiTokenService],
+})
+export class ApiTokenModule {}